feat: 세션 통해 권한 확인 후 페이지 표시여부 결정하는 기능 추가

+++ src/main/java/com/example/sboard/board/controller/BoardController.java
@@ -1,10 +1,14 @@
 package com.example.sboard.board.controller;
 
+import static com.example.sboard.utils.PermissionValidator.validateIsLogin;
+import static com.example.sboard.utils.PermissionValidator.validateIsSelf;
+
 import com.example.sboard.board.domain.Board;
 import com.example.sboard.board.reply.domain.Reply;
 import com.example.sboard.board.reply.service.ReplyService;
 import com.example.sboard.board.service.BoardService;
 import java.util.List;
+import javax.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,7 +19,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequestMapping("/board")
 public class BoardController {
 	private final BoardService boardService;
-
 	private final ReplyService replyService;
 
 	public BoardController(BoardService boardService, ReplyService replyService) {
@@ -45,7 +48,11 @@ public class BoardController {
 	}
 
 	@GetMapping("/insert")
-	public String getBoardInsert(Model model) {
+	public String getBoardInsert(HttpSession session, Model model) {
+		if (!validateIsLogin(session)) {
+			return "redirect:/error/permission";
+		}
+
 		int boardNo = boardService.getLastIndex();
 		model.addAttribute("boardNo", boardNo);
 
@@ -53,28 +60,46 @@ public class BoardController {
 	}
 
 	@PostMapping("/insert")
-	public String postBoardInsert(Board board) {
+	public String postBoardInsert(HttpSession session, Board board) {
+		if (!validateIsLogin(session)) {
+			return "redirect:/error/permission";
+		}
+
 		boardService.insert(board);
 
 		return "redirect:list";
 	}
 
 	@GetMapping("/modify")
-	public String getBoardModify(int boardNo, Model model) {
-		model.addAttribute("board", boardService.get(boardNo));
+	public String getBoardModify(HttpSession session, int boardNo, Model model) {
+		Board board = boardService.get(boardNo);
+		if (!validateIsSelf(session, board.getMemberId())) {
+			return "redirect:/error/permission";
+		}
+
+		model.addAttribute("board", board);
 
 		return "board/modify";
 	}
 
 	@PostMapping("/modify")
-	public String postBoardModify(Board board) {
+	public String postBoardModify(HttpSession session, Board board) {
+		if (!validateIsSelf(session, board.getMemberId())) {
+			return "redirect:/error/permission";
+		}
+
 		boardService.modify(board);
 
 		return "redirect:list";
 	}
 
 	@GetMapping("/delete")
-	public String getBoardDelete(int boardNo) {
+	public String getBoardDelete(HttpSession session, int boardNo) {
+		Board board = boardService.get(boardNo);
+		if (!validateIsSelf(session, board.getMemberId())) {
+			return "redirect:/error/permission";
+		}
+
 		replyService.deleteAllByBoardNo(boardNo);
 		boardService.delete(boardNo);
