+++ src/main/java/com/example/sboard/board/reply/controller/ReplyController.java
@@ -1,7 +1,11 @@
 package com.example.sboard.board.reply.controller;
 
+import static com.example.sboard.utils.PermissionValidator.validateIsLogin;
+import static com.example.sboard.utils.PermissionValidator.validateIsSelf;
+
 import com.example.sboard.board.reply.domain.Reply;
 import com.example.sboard.board.reply.service.ReplyServiceImpl;
+import javax.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -19,7 +23,11 @@ public class ReplyController {
 
 
 	@PostMapping("/insert")
-	public String postReplyInsert(Reply reply) {
+	public String postReplyInsert(HttpSession session, Reply reply) {
+		if (!validateIsLogin(session)) {
+			return "redirect:/error/permission";
+		}
+
 		reply.setReplyNo(replyService.getLastIndex());
 		replyService.insert(reply);
 
@@ -27,8 +35,12 @@ public class ReplyController {
 	}
 
 	@GetMapping("/delete")
-	public String getReplyDelete(int replyNo) {
+	public String getReplyDelete(HttpSession session, int replyNo) {
 		Reply reply = replyService.get(replyNo);
+		if (!validateIsSelf(session, reply.getMemberId())) {
+			return "redirect:/error/permission";
+		}
+
 		replyService.delete(replyNo);
 
 		return "redirect:/board/detail?boardNo=" + reply.getBoardNo();
